Create 11) Maps.go
Demonstrate the use of map in go. It explain the basic of maps and synatx in go. The program also reads a data file that contain the name of the candidates in file and count the votes for each candidates and make a small report from it and tells the winner